feat(api): Add Hilt and KSP dependencies
- Added `com.google.dagger.hilt.android` and `com.google.devtools.ksp` plugins to the app's build.gradle.kts file.
- Added Hilt and Hilts compose dependencies to the app's build.gradle.kts file.
- Created a new class `SlackAssistantApp` in the package `dk.malv.slack.assistant`.
- Moved `SlackAPIClient.kt` to a new package `dk.malv.slack.assistant.di`.
- Renamed `SlackAPIClient.kt` to `AppModule.kt`.
- Updated the code in `AppModule.kt` to use Dagger annotations for dependency injection.
- Updated references to the renamed class in other files.

This commit adds support for Hilt dependency injection framework and KSP (Kotlin Symbol Processing) compiler plugin, allowing for easier management of dependencies in the project.
