+++ app/src/main/java/dk/malv/slack/assistant/SlackAssistantApp.kt
@@ -1,0 +1,7 @@
+package dk.malv.slack.assistant
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class SlackAssistantApp : Application()
